feat: route middleware by role

+++ backend-app/routes/user/user.php
@@ -8,5 +8,5 @@ Route::middleware('auth:api')->group(function () {
     Route::get('/users/{id}', [UserController::class, 'show']);
     Route::post('/users/', [UserController::class,'store']);
     Route::post('/users/{id}', [UserController::class, 'update']);
-    Route::delete('/users/{id}', [UserController::class, 'destroy']);
+    Route::delete('/users/{id}', [UserController::class, 'destroy'])->middleware('role:super_admin');
 });
